Split platform specific time measurement away
We now have a `stopwatch` to measure time between two moments.
